requirements.txt
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.41.1
google-auth-httplib2==0.2.1
google-auth-oauthlib==1.2.3
googleapis-common-protos==1.72.0
httplib2==0.31.0
idna==3.11
ImageHash==4.3.2
MouseInfo==0.1.3
numpy==2.3.4
oauthlib==3.3.1
packaging==25.0
pillow==12.0.0
proto-plus==1.26.1
protobuf==6.33.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==2.0.1
pyobjc-core==12.0
pyobjc-framework-Cocoa==12.0
pyobjc-framework-Quartz==12.0
pyparsing==3.2.5
pyperclip==1.11.0
PyRect==0.2.0
PyScreeze==1.0.1
pytesseract==0.3.13
pytweening==1.2.0
PyWavelets==1.9.0
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
rubicon-objc==0.5.2
scipy==1.16.3
uritemplate==4.2.0
urllib3==2.5.0
